requirements.txt
Django==4.0.7
django-environ==0.9.0
psycopg2-binary==2.9.5
djangorestframework==3.14.0

celery==5.2.7
django-celery-results==2.4.0
django-celery-beat==2.3.0

whitenoise==6.2.0
django-js-asset==2.1.0
django-mptt==0.15.0
django-filter==22.1
django-extensions==3.2.1
django-cors-headers==3.13.0
django-storages==1.13.1

drf-jwt==1.19.2

boto3==1.24.71
attrs==22.1.0

djangorestframework-simplejwt==5.2.2
drf-spectacular==0.27.1

django-redis==5.2.0
gunicorn==20.1.0
sentry-sdk==1.9.8
pyotp==2.9.0
requests[security]